Add shared schedule calendar components for demo/production parity
- Extract shared components from schedule page:
  - types.ts: Common type definitions and API interface
  - utils.ts: Service colors, RRule generation utilities
  - ScheduleCalendarView.tsx: Main FullCalendar component
  - ScheduleDetailDialog.tsx: Schedule detail view
  - ScheduleFormDialog.tsx: Create/edit form with recurrence
  - RecurrenceEditDialog.tsx: Edit scope selection
  - RecurrenceDeleteDialog.tsx: Delete scope selection
- Refactor production /schedule page to use shared components
- Rewrite demo /demo/schedule from table to Google Calendar style
- Use dependency injection pattern for API handlers
- Demo now has full functionality: D&D, resize, create, edit, delete
