+++ web/src/components/schedule/index.ts
@@ -1,0 +1,9 @@
+// Schedule components - shared between production and demo
+
+export * from './types';
+export * from './utils';
+export { ScheduleCalendarView } from './ScheduleCalendarView';
+export { ScheduleDetailDialog } from './ScheduleDetailDialog';
+export { ScheduleFormDialog } from './ScheduleFormDialog';
+export { RecurrenceEditDialog } from './RecurrenceEditDialog';
+export { RecurrenceDeleteDialog } from './RecurrenceDeleteDialog';
